small type changes in cube

+++ random_projection.hpp
@@ -14,9 +14,9 @@ class random_projection{
     const float w;
     const unsigned int k,m,l,new_d;
     f_i ***table_f_i;
-    std::unordered_map<int, my_vector> **hash_table;
+    std::unordered_map<unsigned long int, my_vector> **hash_table;
 
-    int hash_function(my_vector &x, unsigned int &iteration);
+    unsigned long int hash_function(my_vector &x, unsigned int &iteration);
   public:
     random_projection(const unsigned int _l=1, const float _w=4000,//!!! l does not work
               const unsigned int _k=4, const unsigned int _new_d=3,
